lesson 24

+++ app/Http/Controllers/ThreadsController.php
@@ -116,8 +116,9 @@ class ThreadsController extends Controller
      */
     public function destroy($channel, Thread $thread)
     {
-        $thread->delete();
+        $this->authorize('update', $thread);
 
+        $thread->delete();
         if (request()->wantsJson())
             return response([], 204);
